Don't let CMAKE write RPATH
Looks like CMAKE has a default facility which attempts to write the
RPATh of any libraries used upon install.  Since these are hardcoded
anyways in `nix` we can skip this step.  For `Mull` this involved
overwriting `SET` directive in `CmakeLists.txt` because extra steps were
taken to try to force CMAKE to always overwite the RPATH.